atualizacao

+++ ParkingControl.Domain/Shared/ParkingConstants.cs
@@ -2,8 +2,8 @@
 {
     public static class ParkingConstants
     {
-        public const double TOLERANCE_TIME_IN_MINUTES = 10;
-        public const double HALF_HOUR_IN_MINUTES = 30;
-        public const double FULL_HOUR_IN_MINUTES = 60;
+        public const int TOLERANCE_TIME_IN_MINUTES = 10;
+        public const int HALF_HOUR_IN_MINUTES = 30;
+        public const int FULL_HOUR_IN_MINUTES = 60;
     }
 }
